Started creating quest assignment job system (not working yet)
